2018.10.01: * fixed: constants which could not fit in the double representation w/o truncation has converted into fractional number representation with inplace division * changed: merge with the latest library repository changes at commit #e73df2f240b3a7d290903cfdc6bae55777f7f063 from https://github.com/dnwrnr/sgp4/commits/master
git-svn-id: svn+ssh://svn.code.sf.net/p/orbittools/sgp4/trunk@12 96d87c9c-0b25-411f-8f5b-d6b3ddb1d6db

+++ libsgp4/OrbitalElements.cc
@@ -51,7 +51,7 @@ OrbitalElements::OrbitalElements(const Tle& tle)
     const double betao = std::sqrt(betao2);
     const double temp = (1.5 * kCK2) * x3thm1 / (betao * betao2);
     const double del1 = temp / (a1 * a1);
-    const double a0 = a1 * (1.0 - del1 * (1.0 / 3.0 + del1 * (1.0 + del1 * 134.0 / 81.0)));
+    const double a0 = a1 * (1.0 - del1 * (double(1.0) / 3.0 + del1 * (1.0 + del1 * double(134.0) / 81.0)));
     const double del0 = temp / (a0 * a0);
 
     recovered_mean_motion_ = MeanMotion() / (1.0 + del0);
